+++ InformationProtection/functions/Connect-InformationProtection.ps1
@@ -22,6 +22,8 @@
 		Connect-EntraService -TenantID $tenantID -ClientID $clientID -Service AzureRightsManagement
 		Connect-EntraService -TenantID $tenantID -ClientID $clientID -Service MIPSyncService -UseRefreshToken
 
+		Note: This example assumes you previously already imported the InformationProtection module!
+
 		Example 2:
 
 		Connect-EntraService -TenantID $tenantID -ClientID $clientID -Service AzureRightsManagement, MIPSyncService -Certificate $cert
@@ -48,6 +50,23 @@
 	.PARAMETER TenantID
 		The tenant ID of the Entra application to use to authenticate.
 		Defaults to: "organizations" (Which means the tenant, the selected account belongs to)
+
+	.PARAMETER Certificate
+		The certificate to use for authenticating to Entra for the AzureRightsManagement and MIPSyncService services.
+		Requires specifying both ClientID and TenantID.
+		When used, this will first authenticate as application to the provided application, before registering those sessions for the MIP SDK's use.
+
+	.PARAMETER Email
+		Email address to register on the session object.
+		Used for metadata when labelling files.
+
+	.PARAMETER PassThru
+		Returns the MIP session as an object, on top of storing it as the module's default session.
+
+	.PARAMETER LogLevel
+		The level of details to include in the MIP-integrated logging.
+		Logs can be found under "%localappdata%\PowerShell\InformationProtection\logs" or the non-windows equivalent.
+		Defaults to: Error.
 	
 	.EXAMPLE
 		PS C:\> Connect-InformationProtection
@@ -60,6 +79,7 @@
 		Authenticate while creating new EntraAuth sessions for the services "AzureRightsManagement" and "MIPSyncService".
 		This will only use the Authorization Code delegate authentication flow.
 	#>
+	[OutputType([InformationProtection.MipSession])]
 	[CmdletBinding()]
 	param (
 		[hashtable]
@@ -69,25 +89,54 @@
 		$ClientID,
 
 		[string]
-		$TenantID = 'organizations'
+		$TenantID = 'organizations',
+
+		[System.Security.Cryptography.X509Certificates.X509Certificate2]
+		$Certificate,
+
+		[string]
+		$Email,
+
+		[switch]
+		$PassThru,
+
+		[Microsoft.InformationProtection.LogLevel]
+		$LogLevel = 'Error'
 	)
 	begin {
 		$services = $script:_serviceSelector.GetServiceMap($ServiceMap)
 
 		if ($ClientID) {
-			Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.AzureRightsManagement
-			Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.MIPSyncService -UseRefreshToken
+			if (-not $Certificate) {
+				Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.AzureRightsManagement
+				Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.MIPSyncService -UseRefreshToken
+			}
+			else {
+				if (-not ($TenantID -as [guid])) {
+					Stop-PSFFunction -String 'Connect-InformationProtection.Error.NoTenantId' -Cmdlet $PSCmdlet -EnableException $true -Category InvalidArgument
+				}
+				Connect-EntraService -TenantID $TenantID -ClientID $ClientID -Service $services.AzureRightsManagement, $services.MIPSyncService -Certificate $Certificate
+			}
 		}
 
 		Assert-EntraConnection -Cmdlet $PSCmdlet -Service $services.AzureRightsManagement
 		Assert-EntraConnection -Cmdlet $PSCmdlet -Service $services.MIPSyncService
 	}
 	process {
-		$logPath = Join-Path -Path (Get-PSFPath -Name LocalAppData) -ChildPath 'PowerShell\InformationProtection\logs'
-		[InformationProtection.MipHost]::Authenticate(
+		$logPath = Join-Path -Path (Get-PSFPath -Name LocalAppData) -ChildPath "PowerShell\InformationProtection\logs\$([guid]::NewGuid())"
+		$session = [InformationProtection.MipSession]::new()
+		if ($Email) { $session.Email = $Email }
+		$session.Authenticate(
 			(Get-EntraToken -Service $services.AzureRightsManagement),
 			(Get-EntraToken -Service $services.MIPSyncService),
-			$logPath
+			$logPath,
+			$LogLevel
 		)
+		if ($script:_session) {
+			$script:_session.Disconnect()
+		}
+		$script:_session = $session
+
+		if ($PassThru) { $session }
 	}
 }