+++ InformationProtection/functions/Set-MipLabel.ps1
@@ -33,6 +33,16 @@
 		but a file previously labeled under "Privileged" cannot be relabeled as Standard.
 		Defaults to: Privileged (presumably, regular users are not going to be running PowerShell).
 
+	.PARAMETER Session
+		MIP Session to use for the operation.
+		Overrides the use of the default session and would be used in situations when relabeling files from one tenant to another.
+		Use "New-MipSession" to create a standalone session object.
+
+	.PARAMETER Force
+		Forcibly apply the label, even if the file already has the correct label.
+		Will in most cases break the file.
+		Necessary when migrating to the same default builtin label in another tenant.
+
 	.PARAMETER WhatIf
 		If this switch is enabled, no actions are performed but informational messages will be displayed that explain what would happen if the command were to run.
 	
@@ -65,13 +75,22 @@
 		$Justification,
 
 		[Microsoft.InformationProtection.AssignmentMethod]
-		$Method = 'Privileged'
+		$Method = 'Privileged',
+
+		[InformationProtection.MipSession]
+		$Session,
+
+		[switch]
+		$Force
 	)
 	begin {
-		Assert-MIPConnection -Cmdlet $PSCmdlet
+		Assert-MIPConnection -Cmdlet $PSCmdlet -Session $Session
+		$sessionToUse = $script:_session
+		if ($Session.Context) { $sessionToUse = $Session }
+
 		$killIt = $ErrorActionPreference -eq 'Stop'
 
-		$labelObject = Get-MipLabel -Filter $Label
+		$labelObject = Get-MipLabel -Filter $Label -Session $sessionToUse
 		if (-not $labelObject) {
 			Stop-PSFFunction -String 'Set-MipLabel.Error.LabelNotFound' -StringValues $Label -Cmdlet $PSCmdlet -EnableException $true -Category InvalidArgument
 		}
@@ -84,9 +103,19 @@
 	}
 	process {
 		foreach ($filePath in $Path) {
-			$file = [InformationProtection.File]$filePath
+			try { $file = Get-MipFile -Path $filePath -Session $sessionToUse -ErrorAction Stop }
+			catch {
+				Write-Error $_
+				continue
+			}
+			if ($file.LabelID -eq $labelObject.Id -and -not $Force) {
+				Write-PSFMessage -Level Verbose -String 'Set-MipLabel.Already.Labeled' -StringValues $filePath, $labelObject.Name, $labelObject.ID
+				continue
+			}
 			$directory = Split-Path -Path $file.Path
 			$fileName = Split-Path -Path $file.Path -Leaf
+			# Always pick the protected name, because the only files that can accept labels without protection don't switch extensions between protected and unprotected
+			$fileNewName = $file.FileNameProtected
 			$tempNewPath = Join-Path -Path $directory -ChildPath ([Guid]::NewGuid())
 			$tempOldName = [Guid]::NewGuid().ToString()
 			$tempOldPath = Join-Path -Path $directory -ChildPath $tempOldName
@@ -102,7 +131,7 @@
 				}
 
 				# Step 3: Rename labeled file to original name
-				try { Rename-Item -LiteralPath $tempNewPath -NewName $fileName -Force -ErrorAction Stop }
+				try { Rename-Item -LiteralPath $tempNewPath -NewName $fileNewName -Force -ErrorAction Stop }
 				catch {
 					# Rollback and delete new file
 					Rename-Item -LiteralPath $tempOldPath -NewName $fileName -Force