+++ InformationProtection/internal/scripts/preimport.ps1
@@ -10,5 +10,8 @@ After building the module, this file will be completely ignored, adding anything
 
 $moduleRoot = Split-Path (Split-Path $PSScriptRoot)
 
+# Load the Assemblies
+"$moduleRoot\bin\import-dll.ps1"
+
 # Load the strings used in messages
 "$moduleRoot\internal\scripts\strings.ps1"