+++ InformationProtection/changelog.md
@@ -1,5 +1,24 @@
 ﻿# Changelog
 
+## 1.0.17 (2026-04-16)
+
++ New: Get-MipLabel - Returns the currently established MIP session.
++ New: Remove-MipLabel - makes it all go away
++ Upd: Connect-InformationProtection - Add support for a `-Certificate` parameter to simplify application authentication
++ Upd: Connect-InformationProtection - Enables specifying the email address used for metadata on labeled files.
++ Upd: Set-MipLabel - no longer clears files to 0 bytes when applying the same label again.
++ Upd: Set-MipLabel - updates the extension in the same way the desktop client would (.txt -> .ptxt, etc.)
+
+## 1.0.11 (2026-03-12)
+
++ New: Disconnect-InformationProtection - Disconnects from MIP.
++ New: New-MipSession - Creates a new MP Session used to perform file labelling.
++ New: Import-MipSession - Uses a provided MipSession as the default session used by the module.
++ New: Added support for ARM OS-Architectures
++ Upd: Connections are now runspace-specific and no longer process-wide, allowing concurrent connections to different tenants & parallelized operations.
++ Upd: MIP SDK - Update to v1.18.103
++ Fix: Connect-InformationProtection - crashes console when connecting again.
+
 ## 0.9.4 (2025-11-17)
 
 + Upd: Set-MipLabel - added "Method" parameter, to enable Privileged labelling